requirements.txt
easyocr==1.4.1
fastapi==0.70.0
numpy==1.20.2
Pillow==8.4.0
sentence_transformers==2.1.0
torch==1.10.0
transformers==4.12.5
